+++ yosai/core/authc/authc_settings.py
@@ -6,11 +6,22 @@ class AuthenticationSettings:
     """
     def __init__(self, settings):
         self.authc_config = settings.AUTHC_CONFIG
-        self.default_algorithm = self.authc_config.get('default_algorithm',
-                                                       'bcrypt_sha256')
-        self.algorithms = self.authc_config.get('hash_algorithms', None)
+        self.algorithms = self.init_algorithms()
+
+        preferred = self.authc_config.get('preferred_algorithm')
+        self.preferred_algorithm = self.algorithms.get(preferred)
+
         self.account_lock_threshold = self.authc_config.get('account_lock_threshold',
                                                             None)
+        self.totp_context = totp_settings.get('context')
+
+    def init_algorithms(self):
+        algorithms = self.authc_config.get('hash_algorithms', None)
+        if algorithms:
+            return {alg: {"{0}__{1}".format(alg, key): value
+                          for key, value in vals.items()}
+                    for alg, vals in algorithms.items()}
+        return None
 
     def get_config(self, algo):
         """